Timeout settings (#1)
* Fix a bug that created bad parameter names (#60)

When the parameter is called `type` we have to convert this to prevent using the reserved keyword `type` as a variable name. But we were doing it in a few other locations as well. That’s fixed now.

* Fix a bug in the listAffinityGroup call (#61)

It reports the number of VirtualMachines in the groups as being the number of groups found.

* Fix a bug in the createSecurityGroup call (#62) (#63)

* Fix usersecuritygrouplist parameters (#66)

* Fix/improve #66 (#68)

Fixing it this way makes it cleaner to use/set the `usersecuritygrouplist` in clients using this, and keeps it simular in use compared to working with `serviceproviderlist` and `tags`.

Fixes #67

* Fix the authorizeSecurityGroup[In|E]gress responses (#69)

* Alpha version.

* Updated formatting

* blank identifier removed by mistake

* GetTemplateByID zoneid parameter added.

* Region modifications

* Some fixes.